# Optimize Trie Snapshot and Benchmark against deep copy.