--- a/lib/runtime/wasmtime/legacy_imports.go
+++ b/lib/runtime/wasmtime/legacy_imports.go
@@ -110,7 +110,10 @@ func ext_set_storage(c *wasmtime.Caller, keyData, keyLen, valueData, valueLen in
 	key := memory[keyData : keyData+keyLen]
 	val := memory[valueData : valueData+valueLen]
 	logger.Trace("[ext_set_storage]", "key", fmt.Sprintf("0x%x", key), "val", val)
-	err := ctx.Storage.Set(key, val)
+
+	cp := make([]byte, len(val))
+	copy(cp, val)
+	err := ctx.Storage.Set(key, cp)
 	if err != nil {
 		logger.Error("[ext_set_storage]", "error", err)
 		return