<!---

PLEASE READ CAREFULLY

-->

## Changes

<!--

Please provide a brief but specific list of changes made, describe the changes
in functionality rather than the changes in code.

-->

- Implement versioned trie and benchmark.

### Benchmark Results (Snapshot vs Deepcopy)
```
Number of iteration = 1000
```
#### Snapshot
```
=== RUN   TestDeepCopy/Snapshot
2021/02/09 20:09:25 
Alloc = 38 MB
TotalAlloc = 53 MB 
Sys = 71 MB 
NumGC = 4 
2021/02/09 20:09:25 Snapshot to trie took 31.693922ms
```
#### Deepcopy
```
=== RUN   TestDeepCopy/DeepCopy
2021/02/09 20:09:57
Alloc = 4869 MB
TotalAlloc = 4915 MB
Sys = 5093 MB
NumGC = 12
2021/02/09 20:09:57 DeepCopy to trie took 5.219532653s
```

#### Memory (38 MB vs  4869 MB)
#### Execution Time (31.69ms vs  5.21s)

## Tests

<!--

Details on how to run tests relevant to the changes within this pull request.

-->

```
TestSnapshot
TestDeepCopyVsSnapshot
```

## Checklist

<!--

Each empty square brackets below is a checkbox. Replace [ ] with [x] to check
the box after completing the task.

-->

- [x] I have read [CODE_OF_CONDUCT](https://github.com/ChainSafe/gossamer/blob/development/.github/CODE_OF_CONDUCT.md) and [CONTRIBUTING](https://github.com/ChainSafe/gossamer/blob/development/.github/CONTRIBUTING.md) 
- [x] I have provided as much information as possible and necessary
- [x] I have reviewed my own pull request before requesting a review
- [ ] All integration tests and required coverage checks are passing

## Issues

<!--

Please link any issues that this pull request is related to and use the GitHub
supported format for automatically closing issues (ie, closes #123, fixes #123)

See: https://help.github.com/en/articles/closing-issues-using-keywords

-->

- closes #1343