@arijitAD 2021-02-10
> Nice work, On my machine Snapshot is about 150 times faster than DeepCopy, but they both used the same memory:
> 
> ```
> === RUN   TestDeepCopyVsSnapshot/DeepCopy
> 2021/02/09 17:24:27 
> Alloc = 4868 MB 
> TotalAlloc = 4914 MB 
> Sys = 5093 MB 
> NumGC = 12 
> 
> 2021/02/09 17:24:27 DeepCopy to trie took 1.563547599s
>     --- PASS: TestDeepCopyVsSnapshot/DeepCopy (1.56s)
> === RUN   TestDeepCopyVsSnapshot/Snapshot
> 2021/02/09 17:24:27 
> Alloc = 4868 MB 
> TotalAlloc = 4915 MB 
> Sys = 5093 MB 
> NumGC = 12 
> 
> 2021/02/09 17:24:27 DeepCopy to trie took 9.643327ms
> ```

I forgot to trigger garbage collection after each test. It should return the correct stats now.