@arijitAD 2021-02-10
> `TestDeepCopyVsSnapshot` times out:
> ![image](https://user-images.githubusercontent.com/10461622/107404558-12dbac80-6b0f-11eb-85c9-553bd35d204e.png)

<img width="659" alt="Screenshot 2021-02-10 at 5 03 17 PM" src="https://user-images.githubusercontent.com/5843604/107504871-ed5bab00-6bc1-11eb-8b9a-ab76d81ab567.png">
It works for me. Can you remove the timeout and use the above command?